requirements.txt
Pillow==9.5.0
protobuf==3.19.4
streamlit==1.16.0
streamlit-option-menu==0.3.2

